+++ lib/QECCore/src/QECCore.jl
@@ -7,10 +7,8 @@ using Graphs
 
 using DocStringExtensions
 
-using DocStringExtensions
-
 # interfaces
-export distance, parity_matrix, code_n, code_s, code_k, parity_matrix_x, parity_matrix_z, parity_matrix_xz, rate, metacheck_matrix_x, metacheck_matrix_z, metacheck_matrix
+export distance, parity_matrix, code_n, code_s, code_k, parity_matrix_x, parity_matrix_z, rate, metacheck_matrix_x, metacheck_matrix_z, metacheck_matrix
 export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanceAlg
 
 # QEC Codes